Move more to craftassist dialog tests
